rename graph-> metric
