+++ app/components/pool/graphs/pool-graphs.html
@@ -4,61 +4,61 @@
     <div *ngIf="isPaasPool" [other-fix]="reimageFailedNodes" fixMessage="Reimage failed nodes"></div>
 </bl-banner>
 <bl-metrics-monitor class="graphs-container">
-    <bl-metrics-monitor-graph label="Heatmap">
+    <bl-metrics-monitor-metric label="Heatmap">
         <bl-nodes-heatmap blMonitorGraphPreview [pool]="pool" [nodes]="nodes" [showRunningTasks]="false" [showLegend]="false" [interactive]="false"
             [limitNode]="10"></bl-nodes-heatmap>
 
         <bl-nodes-heatmap blMonitorGraphChart [pool]="pool" [nodes]="nodes" [jobs]="jobs"></bl-nodes-heatmap>
-    </bl-metrics-monitor-graph>
-    <bl-metrics-monitor-graph label="Available nodes">
+    </bl-metrics-monitor-metric>
+    <bl-metrics-monitor-metric label="Available nodes">
         <bl-history-graph blMonitorGraphPreview [max]="pool.targetNodes"  [history]="runningNodesHistory.history"
             [interactive]="false" [historySize]="selectedHistoryLength.value">
         </bl-history-graph>
 
         <bl-history-graph blMonitorGraphChart [max]="pool.targetNodes" class="full-display" [history]="runningNodesHistory.history" [historySize]="selectedHistoryLength.value">
         </bl-history-graph>
-    </bl-metrics-monitor-graph>
-    <bl-metrics-monitor-graph label="Running tasks">
+    </bl-metrics-monitor-metric>
+    <bl-metrics-monitor-metric label="Running tasks">
         <bl-history-graph blMonitorGraphPreview [max]="maxRunningTasks" [history]="runningTaskHistory.history" [interactive]="false"
             [historySize]="selectedHistoryLength.value">
         </bl-history-graph>
 
         <bl-history-graph blMonitorGraphChart [max]="maxRunningTasks" class="full-display" [history]="runningTaskHistory.history" [historySize]="selectedHistoryLength.value">
         </bl-history-graph>
-    </bl-metrics-monitor-graph>
+    </bl-metrics-monitor-metric>
     <ng-container *ngIf="appInsightsEnabled">
-        <bl-metrics-monitor-graph label="Cpu">
+        <bl-metrics-monitor-metric label="Cpu">
             <bl-cpu-usage-graph blMonitorGraphPreview #cpuUsage [data]="performanceData" [interactive]="false"></bl-cpu-usage-graph>
             <div blMetricsStatus>{{cpuUsage.status | async}}</div>
 
             <bl-cpu-usage-graph blMonitorGraphChart [data]="performanceData" class="full-display"></bl-cpu-usage-graph>
-        </bl-metrics-monitor-graph>
-        <bl-metrics-monitor-graph label="Memory">
+        </bl-metrics-monitor-metric>
+        <bl-metrics-monitor-metric label="Memory">
             <bl-memory-usage-graph blMonitorGraphPreview #memUsage [data]="performanceData" [interactive]="false"></bl-memory-usage-graph>
             <div blMetricsStatus>{{memUsage.status | async}}</div>
 
             <bl-memory-usage-graph blMonitorGraphChart [data]="performanceData" class="full-display"></bl-memory-usage-graph>
-        </bl-metrics-monitor-graph>
-        <bl-metrics-monitor-graph label="Disk IO">
+        </bl-metrics-monitor-metric>
+        <bl-metrics-monitor-metric label="Disk IO">
             <bl-disk-usage-graph blMonitorGraphPreview #diskUsage [data]="performanceData" [interactive]="false"></bl-disk-usage-graph>
             <div blMetricsStatus>{{diskUsage.status | async}}</div>
 
             <bl-disk-usage-graph blMonitorGraphChart [data]="performanceData" class="full-display"></bl-disk-usage-graph>
-        </bl-metrics-monitor-graph>
-        <bl-metrics-monitor-graph label="Network">
+        </bl-metrics-monitor-metric>
+        <bl-metrics-monitor-metric label="Network">
             <bl-network-usage-graph blMonitorGraphPreview #netUsage [data]="performanceData" [interactive]="false"></bl-network-usage-graph>
             <div blMetricsStatus>{{netUsage.status | async}}</div>
 
             <bl-network-usage-graph blMonitorGraphChart  metric="network" [data]="performanceData" class="full-display"></bl-network-usage-graph>
-        </bl-metrics-monitor-graph>
+        </bl-metrics-monitor-metric>
     </ng-container>
     <ng-container *ngIf="!appInsightsEnabled">
-        <bl-metrics-monitor-graph label="Application Insights">
+        <bl-metrics-monitor-metric label="Application Insights">
             <bl-cpu-usage-graph blMonitorGraphPreview [data]="performanceData" [interactive]="false"></bl-cpu-usage-graph>
             <div blMetricsStatus>Cpu, memory, disk, etc.</div>
 
             <bl-enable-app-insights-doc blMonitorGraphChart></bl-enable-app-insights-doc>
-        </bl-metrics-monitor-graph>
+        </bl-metrics-monitor-metric>
     </ng-container>
 
     <div blMetricsMonitorConfig>
