+++ src/@batch-flask/ui/metrics-monitor/index.ts
@@ -1,3 +1,3 @@
 export * from "./metrics-monitor.component";
-export * from "./metrics-monitor-graph";
+export * from "./metrics-monitor-metric";
 export * from "./metrics-monitor.module";
