--- src/@batch-flask/ui/metrics-monitor/metrics-monitor-graph/index.ts
@@ -1,1 +1,0 @@
-export * from "./metrics-monitor-graph.component";
