+++ src/@batch-flask/ui/metrics-monitor/metrics-monitor-metric/index.ts
@@ -1,0 +1,1 @@
+export * from "./metrics-monitor-metric.component";
