+++ src/@batch-flask/ui/metrics-monitor/metrics-monitor-metric/metrics-monitor-metric.html
